+++ tests/examples/python/helm/__main__.py
@@ -11,12 +11,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from pulumi_kubernetes.core.v1 import Namespace
 from pulumi_kubernetes.helm.v2 import Chart, ChartOpts
+
+namespace = Namespace("test")
 
 values = {"unbound": {"image": {"pullPolicy": "Always"}}}
 
-Chart("unbound", ChartOpts("stable/unbound", values=values))
+Chart("unbound", ChartOpts(
+    "stable/unbound", values=values, namespace=namespace.metadata["name"]))
 
 # Deploy a duplicate chart with a different resource prefix to verify that multiple instances of the Chart
 # can be managed in the same stack.
-Chart("unbound", ChartOpts("stable/unbound", resource_prefix="dup", values=values))
+Chart("unbound", ChartOpts(
+    "stable/unbound", resource_prefix="dup", values=values, namespace=namespace.metadata["name"]))