+++ tests/examples/python/python_test.go
@@ -292,7 +292,7 @@ func TestHelm(t *testing.T) {
 		ExpectRefreshChanges: true, // PodDisruptionBudget status gets updated by the Deployment.
 		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
 			assert.NotNil(t, stackInfo.Deployment)
-			assert.Equal(t, 12, len(stackInfo.Deployment.Resources))
+			assert.Equal(t, 13, len(stackInfo.Deployment.Resources))
 
 			sort.Slice(stackInfo.Deployment.Resources, func(i, j int) bool {
 				ri := stackInfo.Deployment.Resources[i]
@@ -318,11 +318,11 @@ func TestHelm(t *testing.T) {
 			assert.True(t, strings.HasPrefix(pullPolicy.(string), "Always"))
 
 			// Verify the provider resource.
-			provRes := stackInfo.Deployment.Resources[10]
+			provRes := stackInfo.Deployment.Resources[11]
 			assert.True(t, providers.IsProviderType(provRes.URN.Type()))
 
 			// Verify root resource.
-			stackRes := stackInfo.Deployment.Resources[11]
+			stackRes := stackInfo.Deployment.Resources[12]
 			assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
 		},
 	})